feat: refactor WaitingRoomIdGenerator package structure
