+++ src/main/java/com/icebreaker/be/application/waitingroom/WaitingRoomService.java
@@ -3,10 +3,10 @@ package com.icebreaker.be.application.waitingroom;
 import com.icebreaker.be.application.waitingroom.dto.CreateWaitingRoomCommand;
 import com.icebreaker.be.application.waitingroom.dto.WaitingRoomId;
 import com.icebreaker.be.application.waitingroom.event.WaitingRoomEventPublisher;
-import com.icebreaker.be.domain.room.service.WaitingRoomIdGenerator;
 import com.icebreaker.be.domain.user.User;
 import com.icebreaker.be.domain.user.UserRepository;
 import com.icebreaker.be.domain.waitingroom.WaitingRoom;
+import com.icebreaker.be.domain.waitingroom.WaitingRoomIdGenerator;
 import com.icebreaker.be.domain.waitingroom.WaitingRoomParticipant;
 import com.icebreaker.be.domain.waitingroom.WaitingRoomRepository;
 import com.icebreaker.be.domain.waitingroom.WaitingRoomWithParticipants;
